+++ src/drivetrain/holonomic_drive.cpp
@@ -1,19 +1,3 @@
 #include "drivetrain/holonomic_drive.h"
 
-namespace drivetrain {
-// HolonomicDrivetrainMotors GetHolonomicDrivetrainMotors() {
-//   return HolonomicDrivetrainMotors(
-//       hardware::ProsMotorGroup(
-//           constant::kFrontRightMotorPorts, constant::kDrivetrainReverse,
-//           constant::kDrivetrainGearset),
-//       hardware::ProsMotorGroup(
-//           constant::kBackRightMotorPorts, constant::kDrivetrainReverse,
-//           constant::kDrivetrainGearset),
-//       hardware::ProsMotorGroup(
-//           constant::kBackLeftMotorPorts, constant::kDrivetrainReverse,
-//           constant::kDrivetrainGearset),
-//       hardware::ProsMotorGroup(
-//           constant::kFrontLeftMotorPorts, constant::kDrivetrainReverse,
-//           constant::kDrivetrainGearset));
-// }
-}  // namespace drivetrain
+namespace drivetrain {}  // namespace drivetrain
